@@ -1,5 +1,5 @@
 module github.com/dependabot/gomodules-extracted/extract
 
-go 1.15
+go 1.16
 
 require github.com/hmarr/pkgextract v0.1.0
