@@ -25,7 +25,6 @@ var (
 	GOARCH		= envOr("GOARCH", defaultGOARCH)
 	GOOS		= envOr("GOOS", defaultGOOS)
 	GO386		= envOr("GO386", defaultGO386)
-	GOAMD64		= goamd64()
 	GOARM		= goarm()
 	GOMIPS		= gomips()
 	GOMIPS64	= gomips64()
@@ -37,17 +36,16 @@ var (
 
 const (
 	ElfRelocOffset		= 256
-	MachoRelocOffset	= 2048			// reserve enough space for ELF relocations
-	Go115AMD64		= "alignedjumps"	// Should be "alignedjumps" or "normaljumps"; this replaces environment variable introduced in CL 219357.
+	MachoRelocOffset	= 2048	// reserve enough space for ELF relocations
 )
 
-// TODO(1.16): assuming no issues in 1.15 release, remove this and related constant.
-func goamd64() string {
-	return Go115AMD64
-}
-
 func goarm() int {
-	switch v := envOr("GOARM", defaultGOARM); v {
+	def := defaultGOARM
+	if GOOS == "android" && GOARCH == "arm" {
+		// Android arm devices always support GOARM=7.
+		def = "7"
+	}
+	switch v := envOr("GOARM", def); v {
 	case "5":
 		return 5
 	case "6":
@@ -131,11 +129,15 @@ func init() {
 			addexp(f)
 		}
 	}
+
+	// regabi is only supported on amd64.
+	if GOARCH != "amd64" {
+		Regabi_enabled = 0
+	}
 }
 
-func Framepointer_enabled(goos, goarch string) bool {
-	return framepointer_enabled != 0 && (goarch == "amd64" || goarch == "arm64" && (goos == "linux" || goos == "darwin"))
-}
+// Note: must agree with runtime.framepointer_enabled.
+var Framepointer_enabled = GOARCH == "amd64" || GOARCH == "arm64" && (GOOS == "linux" || GOOS == "darwin" || GOOS == "ios")
 
 func addexp(s string) {
 	// Could do general integer parsing here, but the runtime copy doesn't yet.
@@ -159,10 +161,10 @@ func addexp(s string) {
 }
 
 var (
-	framepointer_enabled		int	= 1
 	Fieldtrack_enabled		int
 	Preemptibleloops_enabled	int
 	Staticlockranking_enabled	int
+	Regabi_enabled			int
 )
 
 // Toolchain experiments.
@@ -174,9 +176,9 @@ var exper = []struct {
 	val	*int
 }{
 	{"fieldtrack", &Fieldtrack_enabled},
-	{"framepointer", &framepointer_enabled},
 	{"preemptibleloops", &Preemptibleloops_enabled},
 	{"staticlockranking", &Staticlockranking_enabled},
+	{"regabi", &Regabi_enabled},
 }
 
 var defaultExpstring = Expstring()
