@@ -16,6 +16,8 @@ var OSArchSupportsCgo = map[string]bool{
 	"freebsd/arm":		true,
 	"freebsd/arm64":	true,
 	"illumos/amd64":	true,
+	"ios/amd64":		true,
+	"ios/arm64":		true,
 	"js/wasm":		false,
 	"linux/386":		true,
 	"linux/amd64":		true,
@@ -27,7 +29,7 @@ var OSArchSupportsCgo = map[string]bool{
 	"linux/mipsle":		true,
 	"linux/ppc64":		false,
 	"linux/ppc64le":	true,
-	"linux/riscv64":	false,
+	"linux/riscv64":	true,
 	"linux/s390x":		true,
 	"linux/sparc64":	true,
 	"netbsd/386":		true,
@@ -38,6 +40,7 @@ var OSArchSupportsCgo = map[string]bool{
 	"openbsd/amd64":	true,
 	"openbsd/arm":		true,
 	"openbsd/arm64":	true,
+	"openbsd/mips64":	false,
 	"plan9/386":		false,
 	"plan9/amd64":		false,
 	"plan9/arm":		false,
