@@ -6,11 +6,11 @@ package modfetch
 
 import (
 	"github.com/dependabot/gomodules-extracted/cmd/go/_internal_/cfg"
-	"github.com/dependabot/gomodules-extracted/cmd/go/_internal_/get"
-	"github.com/dependabot/gomodules-extracted/cmd/go/_internal_/str"
+
+	"golang.org/x/mod/module"
 )
 
 // allowInsecure reports whether we are allowed to fetch this path in an insecure manner.
 func allowInsecure(path string) bool {
-	return get.Insecure || str.GlobsMatchPath(cfg.GOINSECURE, path)
+	return cfg.Insecure || module.MatchPrefixPatterns(cfg.GOINSECURE, path)
 }
